Header files created
1.PokemonType Header
2.PokemonChoice Header
3.included in Main cpp
