+++ Pokemon/PokemonChoice.h
@@ -1,0 +1,8 @@
+#pragma once
+enum class PokemonChoice
+{
+    Bulbasaur = 1,
+    Charmander,
+    Squirtle,
+    Pikachu,//default
+};
