+++ Pokemon/PokemonType.h
@@ -1,0 +1,10 @@
+#pragma once
+enum class PokemonType
+{
+    Fire,
+    Electric,
+    Water,
+    Earth,
+    Grass,
+    Normal,//default
+};
